Refactored redux store setup
