+++ src/core/store/dev/enhancers.js
@@ -1,0 +1,4 @@
+import {devToolsEnhancer} from 'redux-devtools-extension';
+import enhancers from '../prod/enhancers';
+
+export default [devToolsEnhancer(), ...enhancers];
