+++ src/core/store/dev/middleware.js
@@ -1,0 +1,3 @@
+import middleware from '../prod/middleware';
+
+export default [...middleware];
