--- src/core/store/devMiddleware.js
@@ -1,3 +1,0 @@
-import middleware from './middleware';
-
-export default [...middleware];
