+++ src/core/store/prod/enhancers.js
@@ -1,0 +1,1 @@
+export default [];
